chatbot-backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "google-generativeai>=0.8.6",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
]
